[inner] classifier
