+++ algorithms/kernel/classifier/classifier_predict.cpp
@@ -35,8 +35,6 @@ namespace prediction
 {
 namespace interface1
 {
-__DAAL_REGISTER_SERIALIZATION_CLASS(Result, SERIALIZATION_CLASSIFIER_PREDICTION_RESULT_ID);
-
 InputIface::InputIface(size_t nElements) : daal::algorithms::Input(nElements) {}
 
 Input::Input() : InputIface(lastModelInputId + 1) {}
@@ -130,49 +128,6 @@ services::Status Input::checkImpl(const daal::algorithms::Parameter *parameter) 
     const size_t predictionDataFeatures = dataTable->getNumberOfColumns();
     DAAL_CHECK_EX(trainingDataFeatures == predictionDataFeatures, services::ErrorIncorrectNumberOfColumns, services::ArgumentName, dataStr());
     return s;
-}
-
-Result::Result() : daal::algorithms::Result(lastResultId + 1) {}
-Result::Result(size_t n) : daal::algorithms::Result(n) {}
-
-/**
- * Returns the prediction result of the classification algorithm
- * \param[in] id   Identifier of the prediction result, \ref ResultId
- * \return         Prediction result that corresponds to the given identifier
- */
-data_management::NumericTablePtr Result::get(ResultId id) const
-{
-    return services::staticPointerCast<data_management::NumericTable, data_management::SerializationIface>(Argument::get(id));
-}
-
-/**
- * Sets the prediction result of the classification algorithm
- * \param[in] id    Identifier of the prediction result, \ref ResultId
- * \param[in] value Pointer to the prediction result
- */
-void Result::set(ResultId id, const data_management::NumericTablePtr &value)
-{
-    Argument::set(id, value);
-}
-
-/**
- * Checks the correctness of the Result object
- * \param[in] input     Pointer to the the input object
- * \param[in] parameter Pointer to the algorithm parameters
- * \param[in] method    Computation method
- */
-services::Status Result::check(const daal::algorithms::Input *input, const daal::algorithms::Parameter *parameter,
-           int method) const
-{
-    return checkImpl(input, parameter);
-}
-
-services::Status Result::checkImpl(const daal::algorithms::Input *input, const daal::algorithms::Parameter *parameter) const
-{
-    size_t nRows = (static_cast<const InputIface *>(input))->getNumberOfRows();
-    data_management::NumericTablePtr resTable = get(prediction);
-
-    return data_management::checkNumericTable(resTable.get(), predictionStr(), data_management::packed_mask, 0, 1, nRows);
 }
 }
