+++ algorithms/kernel/classifier/inner/classifier_predict_fpt_v1.cpp
@@ -1,4 +1,4 @@
-/* file: classifier_predict_fpt.cpp */
+/* file: classifier_predict_fpt_v1.cpp */
 /*******************************************************************************
 * Copyright 2014-2019 Intel Corporation
 *
@@ -34,9 +34,8 @@ namespace prediction
 
 using namespace daal::data_management;
 
-namespace interface2
+namespace interface1
 {
-
 /**
  * Allocates memory for storing prediction results of the classification algorithm
  * \tparam  algorithmFPType     Data type for storing prediction results
@@ -48,26 +47,13 @@ template <typename algorithmFPType>
 DAAL_EXPORT services::Status Result::allocate(const daal::algorithms::Input *input, const daal::algorithms::Parameter *parameter, const int method)
 {
     services::Status st;
-    const Parameter *par = static_cast<const Parameter *>(parameter);
-    DAAL_CHECK(par, services::ErrorNullParameterNotSupported);
-
-    const size_t nRows = (static_cast<const InputIface *>(input))->getNumberOfRows();
-    const size_t nClasses = par->nClasses;
-
-    if(par->resultsToEvaluate & computeClassLabels)
-        set(prediction, HomogenNumericTable<algorithmFPType>::create(1, nRows, NumericTableIface::doAllocate, 0, &st));
-    if(par->resultsToEvaluate & computeClassProbabilities)
-        set(probabilities, HomogenNumericTable<algorithmFPType>::create(nClasses, nRows, NumericTableIface::doAllocate, 0, &st));
-    if(par->resultsToEvaluate & computeClassLogProbabilities)
-        set(logProbabilities, HomogenNumericTable<algorithmFPType>::create(nClasses, nRows, NumericTableIface::doAllocate, 0, &st));
+    set(prediction, HomogenNumericTable<algorithmFPType>::create(1, (static_cast<const InputIface *>(input))->getNumberOfRows(), NumericTableIface::doAllocate, &st));
     return st;
 }
 
-template DAAL_EXPORT services::Status Result::allocate<DAAL_FPTYPE>(const daal::algorithms::Input *input, const daal::algorithms::Parameter *par,
-        const int method);
+template DAAL_EXPORT services::Status Result::allocate<DAAL_FPTYPE>(const daal::algorithms::Input *input, const daal::algorithms::Parameter *par, const int method);
 
 }
-
 }
 }
 }
